feat: add dotenv integration and restructure scottyctl command handling
This change enhances the functionality of the `scottyctl` CLI by integrating the `dotenvy` crate for easier environment variable management. It also restructures the command handling logic to improve readability and maintainability, allowing for a more intuitive interaction with the application.
