+++ scottyctl/src/commands/mod.rs
@@ -1,0 +1,3 @@
+pub mod apps;
+pub mod blueprints;
+pub mod notify;
